% Generated by roxygen2 (4.0.0): do not edit by hand
\name{generate_data}
\alias{generate_data}
\title{Generate new data from a data frame.}
\usage{
generate_data(data, n = 10000, method = "grid")
}
\arguments{
\item{data}{data frame}

\item{n}{desired number of new observations}

\item{method}{method to use, see \code{\link{simvar}}}
}
\description{
This method generates new data that fills the range of
the supplied datasets.
}
\keyword{datagen}

