\name{knnf}
\alias{knnf}
\alias{classify.knnf}
\title{knn, with formula}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
A wrapper function for \code{\link[class]{knn}} to allow use
}
\usage{knnf(formula, data, k=2)}
\arguments{
\item{formula}{classification formula}
\item{data}{training data set}
\item{k}{number of neighbours to use}
}

\details{}

\examples{}
\keyword{classif}
