% Generated by roxygen2 (4.0.0): do not edit by hand
\name{knnf}
\alias{knnf}
\title{A wrapper function for \code{\link[class]{knn}} to allow use
with classifly.}
\usage{
knnf(formula, data, k = 2)
}
\arguments{
\item{formula}{classification formula}

\item{data}{training data set}

\item{k}{number of neighbours to use}
}
\description{
A wrapper function for \code{\link[class]{knn}} to allow use
with classifly.
}
\keyword{classif}

