% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olives.R
\docType{data}
\name{olives}
\alias{olives}
\title{Olives}
\format{
A data frame with 244 rows and 7 variables
}
\description{
The olive oil data consists of the percentage composition of 8 fatty
acids (palmitic, palmitoleic, stearic, oleic, linoleic, linolenic,
arachidic, eicosenoic) found in the lipid fraction of 572 Italian
olive oils. There are 9 collection areas, 4 from southern Italy (North
and South Apulia, Calabria, Sicily), two from Sardinia (Inland and
Coastal) and 3 from northern Italy (Umbria, East and West
Liguria).
}
\references{
Forina, M. and Armanino, C. and Lanteri, S. and
  Tiscornia, E., Classification of olive oils from their fatty acid
  composition, 1983, in Food Research and Data Analysis, edited by
  Martens, H. and Russwurm Jr, H, pages 189-214.
}
