% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.r
\name{variables}
\alias{variables}
\title{Extract predictor and response variables for a model object.}
\usage{
variables(model)
}
\arguments{
\item{model}{model object}
}
\value{
list containing response and predictor variables
}
\description{
Due to the way that most model objects are stored, you
also need to supply the data set you used with the original
data set.  It currently doesn't support models fitted without
using a data argument.
}
\keyword{attribute}
