% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.r
\name{classify}
\alias{classify}
\title{Extract classifications from a variety of methods.}
\usage{
classify(model, data, ...)
}
\arguments{
\item{model}{model object}

\item{data}{data set used in model}

\item{...}{other argument passed on to methods}
}
\description{
If the classification method can produce a matrix of posterior
probabilities (see \code{\link{posterior}}), then that will be used to
calculate the \code{\link{advantage}}.  Otherwise, the classify method
will be used and the advantage calculated using a k-nearest neighbours
approach.
}
\keyword{internal}
