\name{knnf}
\alias{knnf}
\title{A wrapper function for \link[class]{knn} to allow use...}
\usage{knnf(formula, data, k=2)}

\description{
  A wrapper function for \code{\link[class]{knn}} to allow
  use with classifly.
}
\keyword{classif}
\arguments{
  \item{formula}{classification formula}
  \item{data}{training data set}
  \item{k}{number of neighbours to use}
}
