\name{posterior}
\alias{posterior}
\title{Extract posterior group probabilities...}
\usage{posterior(model, data)}

\description{
  Extract posterior group probabilities
}

\details{
  Every classification method seems to provide a slighly
  different way of retrieving the posterior probability of
  group membership.  This function provides a common
  interface to all of them
}
\arguments{
  \item{model}{model object}
  \item{data}{data set used in model}
}
